remove old immediate world rendering
