+++ src/renderer_gl2/r_lightmap.c
@@ -353,59 +353,3 @@ void R_LightMap_EndBuilding()
 	R_LightMap_UploadBlock();
 	R_SelectTextureUnit(TMU_DIFFUSE);
 }
-
-
-
-/*
-=======================
-R_LightMap_EndBuilding
-
-Update lightstyles for surface, pass NULL to disable lightstyles
-=======================
-*/
-void R_LightMap_UpdateLightStylesForSurf(msurface_t *surf)
-{
-	vec3_t lm_colors[MAX_LIGHTMAPS_PER_SURFACE];
-	qboolean hasChanged;
-	int i;
-
-	if (surf == NULL)
-	{
-		VectorSet(lm_colors[0], 1.0f, 1.0f, 1.0f);
-		for (i = 1; i < MAX_LIGHTMAPS_PER_SURFACE; i++)
-			VectorClear(lm_colors[i]);
-		goto change;
-	}
-
-	if (!r_dynamic->value)
-	{
-		for (i = 0; i < MAX_LIGHTMAPS_PER_SURFACE; i++)
-			VectorSet(lm_colors[i], 1.0f, 1.0f, 1.0f);
-	}
-	else
-	{
-		// add light styles
-		for (i = 0; i < MAX_LIGHTMAPS_PER_SURFACE && surf->styles[i] != (MAX_LIGHTSTYLES - 1); i++)
-		{
-			lm_colors[i][0] = r_newrefdef.lightstyles[surf->styles[i]].rgb[0];
-			lm_colors[i][1] = r_newrefdef.lightstyles[surf->styles[i]].rgb[1];
-			lm_colors[i][2] = r_newrefdef.lightstyles[surf->styles[i]].rgb[2];
-		}
-	}
-
-change:
-	hasChanged = false;
-	for (i = 0; i < MAX_LIGHTMAPS_PER_SURFACE; i++)
-	{
-		if (!VectorCompare(lm_colors[i], gl_lms.lm_colors[i]))
-		{
-			VectorCopy(lm_colors[i], gl_lms.lm_colors[i]);
-			hasChanged = true;
-		}
-	}
-
-	if (hasChanged)
-	{
-		R_ProgUniform3fv(LOC_LIGHTSTYLES, MAX_LIGHTMAPS_PER_SURFACE, gl_lms.lm_colors[0]);
-	}
-}
