+++ src/renderer_gl2/renderer_gl2.vcxproj.filters
@@ -40,7 +40,6 @@
     <ClCompile Include="r_model.c" />
     <ClCompile Include="r_main.c" />
     <ClCompile Include="r_misc.c" />
-    <ClCompile Include="r_surf.c" />
     <ClCompile Include="r_warp.c" />
     <ClCompile Include="r_init.c" />
     <ClCompile Include="r_md3.c" />
